requirements.txt
google-adk[database,eval]
yfinance
psutil
litellm>=1.75.5
python-dotenv>=1.0.0
Deprecated
pytest>=8.3.4
fastmcp>=2.3.4
uvicorn>=0.34.0
mypy>=1.15.0
ruff
black
httpx
click>=8.1.8
a2a-sdk>=0.3.4
nest-asyncio
requests>=2.32.4
google-cloud-aiplatform>=1.92.0
google-cloud-storage>=3.0.0
google-genai>=1.14.0
gitpython>=3.1.40
